+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [2.1.0] - 2019-06-14
 - Made RadioButtonCollection and Select countable.
 - Made RadioButtonCollection and Select iterable.
 - Added getLabel and setLabel methods to all form components.
@@ -16,6 +16,7 @@
 ## 1.0.0 - 2018-02-03
 - First stable revision.
 
+[2.1.0]: https://github.com/themichaelhall/bluemvc-forms/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/themichaelhall/bluemvc-forms/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/themichaelhall/bluemvc-forms/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/themichaelhall/bluemvc-forms/compare/v1.0.0...v1.0.1